# generator/tests/openapi.json
{
  "components": {
    "requestBodies": {
...
      "TypeArrayRequest": {
        "content": {
          "application/json": {
            "schema": {
              "properties": {
                "p1": {
                  "items": {
                    "properties": {
                      "Bool": {
                        "properties": {
                          "Bool": {
                            "type": "boolean"
                          }
                        },
                        "type": "object"
                      },
                      "BoolArray": {
                        "items": {
                          "properties": {
                            "Bool": {
                              "type": "boolean"
                            }
                          },
                          "type": "object"
                        },
                        "type": "array"
                      },
                      "BoolArrayArray": {
                        "items": {
                          "items": {
                            "properties": {
                              "Bool": {
                                "type": "boolean"
                              }
                            },
                            "type": "object"
                          },
                          "type": "array"
                        },
                        "type": "array"
                      },
                      "Decimal": {
                        "type": "string"
                      },
                      "DecimalArray": {
                        "items": {
                          "type": "string"
                        },
                        "type": "array"
                      },
                      "DecimalArrayArray": {
                        "items": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        "type": "array"
                      },
                      "Float": {
                        "properties": {
                          "Float32": {
                            "type": "number"
                          },
                          "Float64": {
                            "type": "number"
                          }
                        },
                        "type": "object"
                      },
                      "FloatArray": {
                        "items": {
                          "properties": {
                            "Float32": {
                              "type": "number"
                            },
                            "Float64": {
                              "type": "number"
                            }
                          },
                          "type": "object"
                        },
                        "type": "array"
                      },
                      "FloatArrayArray": {
                        "items": {
                          "items": {
                            "properties": {
                              "Float32": {
                                "type": "number"
                              },
                              "Float64": {
                                "type": "number"
                              }
                            },
                            "type": "object"
                          },
                          "type": "array"
                        },
                        "type": "array"
                      },
                      "Int": {
                        "properties": {
                          "Int": {
                            "type": "integer"
                          },
                          "Int16": {
                            "type": "integer"
                          },
                          "Int16Array": {
                            "items": {
                              "type": "integer"
                            },
                            "type": "array"
                          },
                          "Int16ArrayArray": {
                            "items": {
                              "items": {
                                "type": "integer"
                              },
                              "type": "array"
                            },
                            "type": "array"
                          },
                          "Int32": {
                            "type": "integer"
                          },
                          "Int32Array": {
                            "items": {
                              "type": "integer"
                            },
                            "type": "array"
                          },
                          "Int32ArrayArray": {
                            "items": {
                              "items": {
                                "type": "integer"
                              },
                              "type": "array"
                            },
                            "type": "array"
                          },
                          "Int64": {
                            "type": "integer"
                          },
                          "Int64Array": {
                            "items": {
                              "type": "integer"
                            },
                            "type": "array"
                          },
                          "Int64ArrayArray": {
                            "items": {
                              "items": {
                                "type": "integer"
                              },
                              "type": "array"
                            },
                            "type": "array"
                          },
                          "Int8": {
                            "type": "integer"
                          },
                          "Int8Array": {
                            "items": {
                              "type": "integer"
                            },
                            "type": "array"
                          },
                          "Int8ArrayArray": {
                            "items": {
                              "items": {
                                "type": "integer"
                              },
                              "type": "array"
                            },
                            "type": "array"
                          },
                          "IntArray": {
                            "items": {
                              "type": "integer"
                            },
                            "type": "array"
                          },
                          "IntArrayArray": {
                            "items": {
                              "items": {
                                "type": "integer"
                              },
                              "type": "array"
                            },
                            "type": "array"
                          }
                        },
                        "type": "object"
                      },
                      "IntArray": {
                        "items": {
                          "properties": {
                            "Int": {
                              "type": "integer"
                            },
                            "Int16": {
                              "type": "integer"
                            },
                            "Int16Array": {
                              "items": {
                                "type": "integer"
                              },
                              "type": "array"
                            },
                            "Int16ArrayArray": {
                              "items": {
                                "items": {
                                  "type": "integer"
                                },
                                "type": "array"
                              },
                              "type": "array"
                            },
                            "Int32": {
                              "type": "integer"
                            },
                            "Int32Array": {
                              "items": {
                                "type": "integer"
                              },
                              "type": "array"
                            },
                            "Int32ArrayArray": {
                              "items": {
                                "items": {
                                  "type": "integer"
                                },
                                "type": "array"
                              },
                              "type": "array"
                            },
                            "Int64": {
                              "type": "integer"
                            },
                            "Int64Array": {
                              "items": {
                                "type": "integer"
                              },
                              "type": "array"
                            },
                            "Int64ArrayArray": {
                              "items": {
                                "items": {
                                  "type": "integer"
                                },
                                "type": "array"
                              },
                              "type": "array"
                            },
                            "Int8": {
                              "type": "integer"
                            },
                            "Int8Array": {
                              "items": {
                                "type": "integer"
                              },
                              "type": "array"
                            },
                            "Int8ArrayArray": {
                              "items": {
                                "items": {
                                  "type": "integer"
                                },
                                "type": "array"
                              },
                              "type": "array"
                            },
                            "IntArray": {
                              "items": {
                                "type": "integer"
                              },
                              "type": "array"
                            },
                            "IntArrayArray": {
                              "items": {
                                "items": {
                                  "type": "integer"
                                },
                                "type": "array"
                              },
                              "type": "array"
                            }
                          },
                          "type": "object"
                        },
                        "type": "array"
                      },
                      "IntArrayArray": {
                        "items": {
                          "items": {
                            "properties": {
                              "Int": {
                                "type": "integer"
                              },
                              "Int16": {
                                "type": "integer"
                              },
                              "Int16Array": {
                                "items": {
                                  "type": "integer"
                                },
                                "type": "array"
                              },
                              "Int16ArrayArray": {
                                "items": {
                                  "items": {
                                    "type": "integer"
                                  },
                                  "type": "array"
                                },
                                "type": "array"
                              },
                              "Int32": {
                                "type": "integer"
                              },
                              "Int32Array": {
                                "items": {
                                  "type": "integer"
                                },
                                "type": "array"
                              },
                              "Int32ArrayArray": {
                                "items": {
                                  "items": {
                                    "type": "integer"
                                  },
                                  "type": "array"
                                },
                                "type": "array"
                              },
                              "Int64": {
                                "type": "integer"
                              },
                              "Int64Array": {
                                "items": {
                                  "type": "integer"
                                },
                                "type": "array"
                              },
                              "Int64ArrayArray": {
                                "items": {
                                  "items": {
                                    "type": "integer"
                                  },
                                  "type": "array"
                                },
                                "type": "array"
                              },
                              "Int8": {
                                "type": "integer"
                              },
                              "Int8Array": {
                                "items": {
                                  "type": "integer"
                                },
                                "type": "array"
                              },
                              "Int8ArrayArray": {
                                "items": {
                                  "items": {
                                    "type": "integer"
                                  },
                                  "type": "array"
                                },
                                "type": "array"
                              },
                              "IntArray": {
                                "items": {
                                  "type": "integer"
                                },
                                "type": "array"
                              },
                              "IntArrayArray": {
                                "items": {
                                  "items": {
                                    "type": "integer"
                                  },
                                  "type": "array"
                                },
                                "type": "array"
                              }
                            },
                            "type": "object"
                          },
                          "type": "array"
                        },
                        "type": "array"
                      },
                      "Map": {
                        "type": "object"
                      },
                      "String": {
                        "properties": {
                          "String": {
                            "type": "string"
                          }
                        },
                        "type": "object"
                      },
                      "StringArray": {
                        "items": {
                          "properties": {
                            "String": {
                              "type": "string"
                            }
                          },
                          "type": "object"
                        },
                        "type": "array"
                      },
                      "StringArrayArray": {
                        "items": {
                          "items": {
                            "properties": {
                              "String": {
                                "type": "string"
                              }
                            },
                            "type": "object"
                          },
                          "type": "array"
                        },
                        "type": "array"
                      },
                      "Time": {
                        "type": "string"
                      },
                      "TimeArray": {
                        "items": {
                          "type": "string"
                        },
                        "type": "array"
                      },
                      "TimeArrayArray": {
                        "items": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        "type": "array"
                      },
                      "Uint": {
                        "properties": {
                          "Uint": {
                            "type": "integer"
                          },
                          "Uint16": {
                            "type": "integer"
                          },
                          "Uint32": {
                            "type": "integer"
                          },
                          "Uint64": {
                            "type": "integer"
                          },
                          "Uint8": {
                            "type": "integer"
                          }
                        },
                        "type": "object"
                      },
                      "UintArray": {
                        "items": {
                          "properties": {
                            "Uint": {
                              "type": "integer"
                            },
                            "Uint16": {
                              "type": "integer"
                            },
                            "Uint32": {
                              "type": "integer"
                            },
                            "Uint64": {
                              "type": "integer"
                            },
                            "Uint8": {
                              "type": "integer"
                            }
                          },
                          "type": "object"
                        },
                        "type": "array"
                      },
                      "UintArrayArray": {
                        "items": {
                          "items": {
                            "properties": {
                              "Uint": {
                                "type": "integer"
                              },
                              "Uint16": {
                                "type": "integer"
                              },
                              "Uint32": {
                                "type": "integer"
                              },
                              "Uint64": {
                                "type": "integer"
                              },
                              "Uint8": {
                                "type": "integer"
                              }
                            },
                            "type": "object"
                          },
                          "type": "array"
                        },
                        "type": "array"
                      }
                    },
                    "type": "object"
                  },
                  "type": "array"
                }
              },
              "type": "object"
            }
          }
        }
      },
...
    },
...
  },
  "info": {
    "title": "",
    "version": ""
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/TypeArray": {
      "post": {
        "operationId": "TypeArray",
        "requestBody": {
          "$ref": "#/components/requestBodies/TypeArrayRequest"
        },
        "responses": {
          "200": {
            "description": "success"
          }
        }
      }
    },
...
  }
}